refactor: Extract P2P & signing patterns (Phase 2, #398)
Extract 5 duplicated patterns from FFI/CLI into shared crates:

- parse_multiaddr_with_peer_id → p2p::address (Extraction 7)
- poll_until_connected → P2PHostHandle method (Extraction 8)
- resolve_peer_addresses → P2PHostHandle method (Extraction 9)
- retry_doc → db::push_docs (Extraction 10)
- resolve_signing_config fix in txn/exec.rs (Extraction 11)

Net: -224 lines, 1 file created, 10 files modified.
