+++ crates/p2p/src/address.rs
@@ -1,0 +1,35 @@
+/// Parsed multiaddr containing peer ID and transport address.
+#[derive(Debug, Clone)]
+pub struct ParsedMultiaddr {
+    pub peer_id: libp2p::PeerId,
+    pub transport_addr: libp2p::Multiaddr,
+}
+
+/// Parse a full multiaddr string that includes a peer ID.
+///
+/// Expects format like: `/ip4/127.0.0.1/tcp/9171/p2p/12D3KooW...`
+///
+/// Returns the peer ID and the transport address (without /p2p component).
+pub fn parse_multiaddr_with_peer_id(addr_str: &str) -> Result<ParsedMultiaddr, String> {
+    let full_addr: libp2p::Multiaddr = addr_str
+        .parse()
+        .map_err(|e| format!("invalid multiaddr '{}': {}", addr_str, e))?;
+
+    let peer_id = full_addr
+        .iter()
+        .find_map(|p| match p {
+            libp2p::multiaddr::Protocol::P2p(id) => Some(id),
+            _ => None,
+        })
+        .ok_or_else(|| format!("multiaddr '{}' does not contain peer ID", addr_str))?;
+
+    let transport_addr: libp2p::Multiaddr = full_addr
+        .iter()
+        .filter(|p| !matches!(p, libp2p::multiaddr::Protocol::P2p(_)))
+        .collect();
+
+    Ok(ParsedMultiaddr {
+        peer_id,
+        transport_addr,
+    })
+}
